moved VolumeMeter to a seperate component
